Major updates

+++ dashboard_new.py
@@ -4,6 +4,7 @@ from st_aggrid import AgGrid, GridOptionsBuilder
 import pandas as pd
 import streamlit as st
 import plotly.express as px
+import open3d as o3d
 import requests
 import re
 import os
@@ -14,6 +15,7 @@ import json
 import fetch_ipc as fipc
 import streamlit.components.v1 as components
 import numpy as np
+import traceback
 
 
 @st.cache_data
@@ -181,7 +183,7 @@ def download_file(remote_path, local_folder, local_name):
                     return f"{local_folder}/{extracted_items[1]}"
             return f"{local_folder}/{local_name}.{file_extn}"
     except Exception as e:
-        print(e)
+        print(traceback.format_exc())
         return ""
 
 
@@ -189,18 +191,9 @@ def download_file(remote_path, local_folder, local_name):
 def get_plant_detection_csv(_session, index, season, sensor, crop, date):
     if re.search("3d", sensor, re.IGNORECASE):
         # 3D sensors don't have Plant Detection CSVs, so we find approx RGB date
-        dates_RGB = get_date_list(_session, index, season, "stereoTop", crop, "0")
-        date_RGB = ""
-        date_3D_obj = datetime.strptime(date, "%Y-%m-%d")
-        for date_id in dates_RGB.keys():
-            dte_obj = datetime.strptime(date_id, "%Y-%m-%d")
-            # to check neighboring dates in case of no exact match
-            if (date_3D_obj == dte_obj) or (
-                date_3D_obj == dte_obj + timedelta(days=1)
-                or date_3D_obj == dte_obj - timedelta(days=1)
-            ):
-                date_RGB = date_id
-                break
+        date_RGB = get_closest_date(
+            _session, index, season, "stereoTop", crop, "0", date
+        )
         if date_RGB == "":
             st.write(
                 f"No Plant Detection CSV is present for this sensor on this date. ({date})"
@@ -247,9 +240,25 @@ def get_plant_detection_csv(_session, index, season, sensor, crop, date):
                         print("Plant Detection CSV downloaded")
                         return local_file
                     except Exception as e:
-                        print(e)
+                        print(traceback.format_exc())
                         return ""
     return ""
+
+
+def get_closest_date(_session, index, season, sensor, crop, level, date):
+    dates = get_date_list(_session, index, season, sensor, crop, level)
+    closest_date = ""
+    date_3D_obj = datetime.strptime(date, "%Y-%m-%d")
+    for date_id in dates.keys():
+        dte_obj = datetime.strptime(date_id, "%Y-%m-%d")
+        # to check neighboring dates in case of no exact match
+        if (date_3D_obj == dte_obj) or (
+            date_3D_obj == dte_obj + timedelta(days=1)
+            or date_3D_obj == dte_obj - timedelta(days=1)
+        ):
+            closest_date = date_id
+            break
+    return closest_date
 
 
 @st.cache_data
@@ -262,7 +271,7 @@ def download_fieldbook(index, season):
         print(f"Fieldbook downloaded.")
         return local_file
     except Exception as e:
-        print(e)
+        print(traceback.format_exc())
         return ""
 
 
@@ -279,14 +288,30 @@ def data_analysis(
     # make field book dataframe based on its extension
     if field_book_name.split(".")[1] == "xlsx":
         try:
-            field_book_df = pd.read_excel(
+            all_sheets = pd.read_excel(
                 io=field_book_name,
-                sheet_name=lambda sheet: "field" in sheet.lower()
-                and "book" in sheet.lower(),
+                sheet_name=None,
                 engine="openpyxl",
             )
+
+            # Filter sheets based on condition
+            filtered_sheets = [
+                sheet
+                for sheet in all_sheets.keys()
+                if "field" in sheet.lower() and "book" in sheet.lower()
+            ]
+
+            # Read the desired sheet
+            field_book_df = pd.read_excel(
+                io=field_book_name,
+                sheet_name=filtered_sheets[
+                    0
+                ],  # Use the first sheet that satisfies the condition
+                engine="openpyxl",
+            )
+
         except Exception as e:
-            print(e)
+            print(traceback.format_exc())
             return
     elif field_book_name.split(".")[1] == "csv":
         field_book_df = pd.read_csv(field_book_name)
@@ -314,8 +339,7 @@ def data_analysis(
                 axis=1,
                 inplace=True,
             )
-            st.dataframe(result)
-            # file_fetcher = create_file_fetcher(_session, season, date, crop)
+            file_fetcher = create_file_fetcher(_session, index, season, date, crop)
             # create_filter(file_fetcher, combined_data=result, sensor=sensor)
 
         else:
@@ -354,7 +378,7 @@ def data_analysis(
             )
     except Exception as e:
         # result dataframe is empty or has not been created, allow the users to download the plant detection csv and fieldbook csv
-        print(e)
+        print(traceback.format_exc())
         st.subheader("Some Error occured (might be a merge issue).")
         st.write(
             "Below are the links to download the Plant Detection and the Fieldbook CSV"
@@ -374,6 +398,9 @@ def data_analysis(
 
 
 def extra_files(_session, combined_df, index, season, crop, sensor, date):
+    combined_df[["lat", "lon"]] = (
+        combined_df[["lat", "lon"]].apply(pd.to_numeric).apply(lambda x: round(x, 7))
+    )
     if "3D" in sensor:
         download_loc = dload_indv_plant_data_3D(_session, index, season, crop, date)
         if download_loc != "":
@@ -383,10 +410,15 @@ def extra_files(_session, combined_df, index, season, crop, sensor, date):
             plant_clustering_path = download_plant_clustering_csv(
                 index, season, "stereoTop"
             )
-            if download_loc != "":
-                plant_clustering_df = pd.read_csv(download_loc).loc[
+            if plant_clustering_path != "":
+                plant_clustering_df = pd.read_csv(plant_clustering_path).loc[
                     :, ["plant_name", "lat", "lon"]
                 ]
+                plant_clustering_df[["lat", "lon"]] = (
+                    plant_clustering_df[["lat", "lon"]]
+                    .apply(pd.to_numeric)
+                    .apply(lambda x: round(x, 7))
+                )
                 ind_plant_df = pd.read_csv(plant_data_path)
                 # Taking very long time - 3 mins (try merging plant_clustering with ind first)
                 ind_plant_df = ind_plant_df.merge(plant_clustering_df, on="plant_name")
@@ -401,31 +433,45 @@ def extra_files(_session, combined_df, index, season, crop, sensor, date):
                     plant_clustering_df = pd.read_csv(path).loc[
                         :, ["plant_name", "plot", "genotype", "lat", "lon"]
                     ]
+                    plant_clustering_df[["lat", "lon"]] = (
+                        plant_clustering_df[["lat", "lon"]]
+                        .apply(pd.to_numeric)
+                        .apply(lambda x: round(x, 7))
+                    )
                     combined_df = combined_df.merge(
                         plant_clustering_df, on=["plot", "genotype"]
                     )
                     return combined_df
-                except:
+                except Exception as e:
+                    print(traceback.format_exc())
                     return pd.DataFrame()
-            else:
-                path = download_plant_clustering_csv(index, season, sensor)
-                if path != "":
-                    try:
-                        plant_clustering_df = pd.read_csv(path).loc[
-                            :, ["plant_name", "lat", "lon"]
-                        ]
-                        combined_df = combined_df.merge(
-                            plant_clustering_df, on=["lat", "lon"]
-                        )
-                        return combined_df
-                    except:
-                        return pd.DataFrame()
+        else:
+            path = download_plant_clustering_csv(index, season, sensor)
+            if path != "":
+                try:
+                    plant_clustering_df = pd.read_csv(path).loc[
+                        :,
+                        ["plant_name", "lat", "lon"],
+                    ]
+                    plant_clustering_df[["lat", "lon"]] = (
+                        plant_clustering_df[["lat", "lon"]]
+                        .apply(pd.to_numeric)
+                        .apply(lambda x: round(x, 7))
+                    )
+                    combined_df = combined_df.merge(
+                        plant_clustering_df, on=["lat", "lon"]
+                    )
+                    return combined_df
+                except:
+                    print(traceback.format_exc())
+                    return pd.DataFrame()
     return pd.DataFrame()
 
 
 def dload_indv_plant_data_3D(_session, index, season, crop, date):
     try:
         path = index[season.split(" ")[1]]["metadata"]["volume-entropy"][crop]
+        print(path)
         sublist = _session.list(path)
         date_found = False
         for file in sublist:
@@ -448,7 +494,7 @@ def dload_indv_plant_data_3D(_session, index, season, crop, date):
         st.write(f"Couldn't find the necessary file for plant vis. for this date")
         return ""
     except Exception as e:
-        print(e)
+        print(traceback.format_exc())
         st.write(
             f"Problem occurered while downloading necessary files for 3D sensor. Contact Phytooracle Staff"
         )
@@ -456,31 +502,40 @@ def dload_indv_plant_data_3D(_session, index, season, crop, date):
 
 
 def combine_csv_into_one(folder, csv_name):
+    if os.path.exists(f"indv_plant_data/{csv_name}.csv"):
+        return f"indv_plant_data/{csv_name}.csv"
     try:
         all_files = glob.glob(folder + "/*.csv")
         li = []
         for filename in all_files:
             # to avoid other date csv
-            pattern = r"\w+_\d{4}-\d{2}-\d{2}_\d{2}"
+            pattern = r"\w+_\d{4}-\d{2}-\d{2}_\d{2}.csv"
             if not re.match(pattern, filename):
                 df = pd.read_csv(filename, index_col=None, header=0)
                 li.append(df)
-        result_frame = pd.concat(li, axis=0, ignore_index=True)
-        result_frame.to_csv(f"indv_plant_data/{csv_name}.csv", index=False)
-        shutil.rmtree(folder)
-        return f"indv_plant_data/{csv_name}.csv"
+        if li != []:
+            result_frame = pd.concat(li, axis=0, ignore_index=True)
+            result_frame.to_csv(f"indv_plant_data/{csv_name}.csv", index=False)
+            shutil.rmtree(folder)
+            return f"indv_plant_data/{csv_name}.csv"
+        return ""
     except:
+        print(traceback.format_exc())
         return ""
 
 
 def download_plant_clustering_csv(index, season, sensor):
+    if os.path.exists(
+        f"plant_clustering/season_{season.split(' ')[1]}_{sensor}_pc.csv"
+    ):
+        return f"plant_clustering/season_{season.split(' ')[1]}_{sensor}_pc.csv"
     try:
         path = index[season.split(" ")[1]]["metadata"]["plant-clustering"][sensor]
         return download_file(
             path, "plant_clustering", f"season_{season.split(' ')[1]}_{sensor}_pc"
         )
     except Exception as e:
-        print(e)
+        print(traceback.format_exc())
         return ""
 
 
@@ -490,7 +545,179 @@ def convert_df(df):
     return df.to_csv(index=False).encode("utf-8")
 
 
-# def create_file_fetcher(_session, season, date, crop):
+def create_file_fetcher(_session, index, season, date, crop):
+    # date issue. incomplete
+    closest_date = get_closest_date(_session, index, season, "scanner3DTop", crop, date)
+    if closest_date is None:
+        # There are no level 2 point clouds for the chosen date or any date near it
+        file_fetcher = None
+    else:
+        try:
+            level_2_path = index[season]["paths"]["2"][crop]
+            cyverse_path = f"{level_2_path}/{closest_date}/individual_plants_out/{closest_date}_segmentation_pointclouds_index"
+            local_idx_path = download_file(
+                cyverse_path,
+                "visualization",
+                f"{closest_date}_segmentation_pointclouds_index.txt",
+            )
+            file_fetcher = fipc.Fetcher(
+                "individually_called_point_clouds",
+                season,
+                "level_2",
+                closest_date,
+                crop,
+                local_idx_path,
+            )
+        except:
+            print(traceback.format_exc())
+            return None
+    return file_fetcher
+
+
+def create_filter(file_fetcher, combined_data, sensor):
+    """Creates a dynamic fiter
+
+    Args:
+      - combined_data (pandas df): Everything in this dataframe
+      - sensor (string): selected sensor
+    """
+    filter_options = []
+    pn_exists = False
+    extra = {
+        "alwaysShowVerticalScroll": True,
+        "alwaysShowHorizontalScroll": True,
+        "autoHeight": True,
+    }
+
+    for column_name in combined_data.columns:
+        if column_name == "plant_name":
+            pn_exists = True
+        if not re.search(f"lon|lat|max|min|date", column_name, re.IGNORECASE):
+            filter_options.append(column_name)
+    selected_column_name = filter_sec.selectbox(
+        "Choose an Attribute", sorted(filter_options)
+    )
+    col1.header("All Data")
+    all_gb = GridOptionsBuilder.from_dataframe(combined_data)
+    # configure column definitions
+    for column_name in combined_data.columns:
+        all_gb.configure_column(column_name, filter=True)
+    all_gb.configure_grid_options(**extra)
+    gridOptions = all_gb.build()
+    with col1:
+        selected = AgGrid(
+            combined_data,
+            gridOptions=gridOptions,
+            theme="balham",
+            height=450,
+        )
+    selected_columns = []
+    exact_column_name = selected_column_name
+    for column_name in combined_data.columns:
+        if re.search(
+            f"{selected_column_name}|lon|lat|max|min|name|date",
+            column_name,
+            re.IGNORECASE,
+        ):
+            if re.search(selected_column_name, column_name, re.IGNORECASE):
+                exact_column_name = column_name
+            selected_columns.append(column_name)
+    filtered_df = combined_data.loc[:, combined_data.columns.isin(selected_columns)]
+
+    if selected_column_name == "genotype":
+        selected_genotype = col2.selectbox(
+            "Genotype", filtered_df["genotype"].unique(), key="gf_date_one"
+        )
+
+        filtered_df = filtered_df[filtered_df["genotype"].isin([selected_genotype])]
+
+    if selected_column_name == "range":
+        selected_range = col2.select_slider(
+            "Select a range: ", options=filtered_df["range"].unique()
+        )
+
+        filtered_df = filtered_df[filtered_df["range"].isin([selected_range])]
+
+    if selected_column_name == "plot":
+        selected_plot = col2.select_slider(
+            "Select a plot: ", options=filtered_df["plot"].unique()
+        )
+
+        filtered_df = filtered_df[filtered_df["plot"].isin([selected_plot])]
+
+    col2.header("Filtered Data")
+    filtered_gb = GridOptionsBuilder.from_dataframe(filtered_df)
+    # configure column definitions
+    for column_name in filtered_df.columns:
+        filtered_gb.configure_column(column_name, filter=True)
+    filtered_gb.configure_grid_options(**extra)
+    filtered_gb.configure_selection(selection_mode="single", use_checkbox=True)
+    gridOptions = filtered_gb.build()
+
+    # set aggrid table to column two and watch for events
+    with col2:
+        selected = AgGrid(
+            filtered_df,
+            gridOptions=gridOptions,
+            theme="balham",
+            height=450,
+        )  # get which row user selects of the
+
+        # vizualization on point clouds is possible and a plant was selected use callback
+        if selected["selected_rows"]:
+            callback(file_fetcher, selected["selected_rows"][0]["plant_name"])
+
+    col1.download_button(
+        label="Download All Data",
+        data=convert_df(combined_data),
+        file_name=f"{combined_data.iloc[0, 0]}_combined_data.csv",
+        mime="text/csv",
+    )
+    col2.download_button(
+        label="Download Filtered Data (Co-ordinates + Selected Field)",
+        data=convert_df(filtered_df),
+        file_name=f"{combined_data.iloc[0, 0]}_filtered_data.csv",
+        mime="text/csv",
+    )
+
+
+def callback(file_fetcher, crop_name):
+    """
+    Callback function to download 3D data of a plant, read a point cloud file, apply an offset to the x and y coordinates,
+    create a DataFrame from the point cloud data, and generate a 3D scatter plot of the point cloud using Plotly.
+    """
+    # make change to accomodate filters bc table is not file
+    plant_3d_data = file_fetcher.download_plant_by_index(crop_name)
+
+    # return value is not correct path so get correct path
+    path_final_file = f"individually_called_point_clouds/{crop_name}_timeseries/{file_fetcher.date}_final.ply"
+
+    pcd = o3d.io.read_point_cloud(path_final_file)
+    # Apply offset after opening the point cloud
+    x_offset = 409000
+    y_offset = 3660000
+
+    #   adjust based on offsets
+    points = np.asarray(pcd.points)
+    xs = points[:, 0] - x_offset
+    ys = points[:, 1] - y_offset
+    zs = points[:, 2]
+    df_dict = {"x": xs, "y": ys, "z": zs}
+    df = pd.DataFrame(df_dict)
+    # Use plotly to display stuff
+    color_scale = [[0.0, "yellow"], [1.0, "green"]]
+    fig = px.scatter_3d(
+        df,
+        title=crop_name,
+        x="x",
+        y="y",
+        z="z",
+        color="z",
+        color_continuous_scale=color_scale,
+    )
+
+    fig.update_traces(marker=dict(size=3))
+    dist_col.plotly_chart(fig, use_container_width=True)
 
 
 def main():
@@ -546,11 +773,6 @@ def main():
             selected_date = st.sidebar.select_slider(
                 "Select a date:", list(level_0_dates.keys())
             )
-            date_two = st.sidebar.checkbox("Select another date")
-            if date_two:
-                selected_date_two = st.sidebar.select_slider(
-                    "Select a second date: ", list(level_0_dates.keys())
-                )
             # for organizing stuff on the screen
             global col1, col2, filter_sec, vis_container, plotly_col, dist_col
             filter_sec = st.container()
